Create dijkstra_algorithm.py
Write a (Python, C++, or Java) program that implements Dijkstra's algorithm. There are many solutions for this online. As a result, you will need to come to office hours and do a code-walkthrough, explaining your code to me, in order to receive appropriate credit for your work. 

Your program will accept cost matrix and paths as input.